refactor: move skills tool to tools/ directory
Reorganize code structure for better scalability:

**New Structure:**
```
src/tools/              # Agent tools (future-ready)
  ├── index.ts         # Tools index with documentation
  └── skills.ts        # Skills tool implementation

src/skills/            # Skills system (unchanged)
  ├── provider.ts      # Skills provider
  ├── importer.ts      # Git importer
  ├── import-spec.ts   # Import spec parser
  └── index.ts         # Skills exports
```

**Changes:**
- Move createSkillsTool from src/skills/tool.ts to src/tools/skills.ts
- Create src/tools/index.ts with documentation for future tools
- Update import paths in src/skills/index.ts
- All 143 tests passing ✓

**Benefits:**
- Clear separation: tools/ for tool implementations, skills/ for domain logic
- Scalable structure for future tools (read, write, bash, git, todo, etc.)
- Better organization following single responsibility principle
